refactor(frontend): adjust baodaozai types
