--- packages/frontend/src/services/call-baodaozai-test/types.ts
@@ -1,15 +1,0 @@
-export type GeneralBaodaozaiState =
-  | 'default'
-  | 'idel-sleep'
-  | 'idel-read'
-  | 'idel-enlighten'
-  | 'dialog-speaker'
-  | 'dialog-read'
-  | 'dialog-enlighten'
-
-export type QaBaodaozaiState =
-  | 'default'
-  | 'enlighten-send'
-  | 'enlighten-fault'
-  | 'enlighten-correct'
-  | 'enlighten-ask'
