+++ packages/frontend/src/services/call-baodaozai/types.ts
@@ -32,3 +32,19 @@ export type BaodaozaiChoiceQuestion = {
 export type BaodaozaiQuestion = BaodaozaiEssayQuestion | BaodaozaiChoiceQuestion
 
 export type BaodaozaiQuestions = BaodaozaiQuestion[]
+
+export type GeneralBaodaozaiState =
+  | 'default'
+  | 'idel-sleep'
+  | 'idel-read'
+  | 'idel-enlighten'
+  | 'dialog-speaker'
+  | 'dialog-read'
+  | 'dialog-enlighten'
+
+export type QaBaodaozaiState =
+  | 'default'
+  | 'enlighten-send'
+  | 'enlighten-fault'
+  | 'enlighten-correct'
+  | 'enlighten-ask'
